+++ server/src/main/java/org/opensearch/transport/InboundPipeline.java
@@ -38,6 +38,8 @@ import org.opensearch.common.lease.Releasable;
 import org.opensearch.common.lease.Releasables;
 import org.opensearch.common.util.PageCacheRecycler;
 import org.opensearch.core.common.breaker.CircuitBreaker;
+import org.opensearch.transport.nativeprotocol.InboundAggregator;
+import org.opensearch.transport.nativeprotocol.InboundDecoder;
 import org.opensearch.transport.nativeprotocol.NativeInboundBytesHandler;
 
 import java.io.IOException;
