+++ server/src/main/java/org/opensearch/transport/ProtocolMessageHandler.java
@@ -19,6 +19,11 @@ public interface ProtocolMessageHandler {
 
     /**
      * Handles the message received on the channel.
+     * @param channel the channel on which the message was received
+     * @param message the message received
+     * @param startTime the start time
+     * @param slowLogThresholdMs the threshold for slow logs
+     * @param messageListener the message listener
      */
     public void messageReceived(
         TcpChannel channel,
@@ -30,7 +35,7 @@ public interface ProtocolMessageHandler {
 
     /**
      * Sets the message listener to be used by the handler.
-     * @param listener
+     * @param listener the message listener
      */
     public void setMessageListener(TransportMessageListener listener);
 }
