+++ server/src/main/java/org/opensearch/transport/nativeprotocol/NativeInboundBytesHandler.java
@@ -13,9 +13,7 @@ import org.opensearch.common.lease.Releasable;
 import org.opensearch.common.lease.Releasables;
 import org.opensearch.core.common.bytes.CompositeBytesReference;
 import org.opensearch.transport.Header;
-import org.opensearch.transport.InboundAggregator;
 import org.opensearch.transport.InboundBytesHandler;
-import org.opensearch.transport.InboundDecoder;
 import org.opensearch.transport.ProtocolInboundMessage;
 import org.opensearch.transport.StatsTracker;
 import org.opensearch.transport.TcpChannel;
